requirements.txt
websocket-client>=1.6.0
fastapi>=0.110.0
uvicorn>=0.29.0
websockets>=12.0
